requirements.txt
anyio==3.6.2
attrs==21.4.0
blessed==1.20.0
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.2
distlib==0.3.6
exceptiongroup==1.1.1
filelock==3.10.6
gitdb==4.0.10
GitPython==3.1.41
grpcio==1.59.2
grpcio-tools==1.59.2
h11==0.14.0
h2o-wave==1.3.0
httpcore==0.16.3
httpx==0.23.3
identify==2.5.22
idna==3.4
importlib-metadata==6.0.0
influxdb==5.3.1
iniconfig==2.0.0
inquirer==2.10.1
markdown-it-py==2.2.0
mdurl==0.1.2
msgpack==1.0.5
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.21.6
pandas==2.0.3
pbr==5.11.1
platformdirs==3.2.0
protobuf==4.25.1
pycparser==2.21
Pygments==2.15.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2022.7.1
PyYAML==6.0
qtoml==0.3.1
readchar==4.0.5
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.3.3
setuptools==65.5.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
stevedore==3.5.2
stringcase==1.2.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
virtualenv==20.21.0
wcwidth==0.2.6
zipp==3.15.0
